requirements.txt
PyYAML
pygame_cards
rich
types-PyYAML
